+++ example/lib/swagger_generated_code/pet_service_json.swagger.dart
@@ -1,14 +1,16 @@
 // ignore_for_file: type=lint
 
 import 'package:json_annotation/json_annotation.dart';
+import 'package:json_annotation/json_annotation.dart' as json;
 import 'package:collection/collection.dart';
 import 'dart:convert';
 
-import 'overriden_models.dart';
 import 'package:chopper/chopper.dart';
 
 import 'client_mapping.dart';
 import 'dart:async';
+import 'package:http/http.dart' as http;
+import 'package:http/http.dart' show MultipartFile;
 import 'package:chopper/chopper.dart' as chopper;
 import 'pet_service_json.enums.swagger.dart' as enums;
 export 'pet_service_json.enums.swagger.dart';
@@ -22,20 +24,28 @@ part 'pet_service_json.swagger.g.dart';
 
 @ChopperApi()
 abstract class PetServiceJson extends ChopperService {
+  static Future<void> Function(Future<void> Function())? _handleConnection;
+
   static PetServiceJson create({
     ChopperClient? client,
+    http.Client? httpClient,
     Authenticator? authenticator,
+    Converter? converter,
     Uri? baseUrl,
     Iterable<dynamic>? interceptors,
+    Future<void> Function(Future<void> Function())? handleConnection,
   }) {
+    _handleConnection = handleConnection;
+
     if (client != null) {
       return _$PetServiceJson(client);
     }
 
     final newClient = ChopperClient(
         services: [_$PetServiceJson()],
-        converter: $JsonSerializableConverter(),
+        converter: converter ?? $JsonSerializableConverter(),
         interceptors: interceptors ?? [],
+        client: httpClient,
         authenticator: authenticator,
         baseUrl: baseUrl ?? Uri.parse('http://petstore.swagger.io/v2'));
     return _$PetServiceJson(newClient);
@@ -46,15 +56,23 @@ abstract class PetServiceJson extends ChopperService {
   Future<chopper.Response> petPost({
     required Pet? body,
     String? cacheControl,
-  }) {
+  }) async {
     generatedMapping.putIfAbsent(Pet, () => Pet.fromJsonFactory);
 
+    if (_handleConnection != null) {
+      Future<chopper.Response>? response;
+      await _handleConnection!(() async {
+        response = _petPost(body: body, cacheControl: cacheControl);
+        await response;
+      });
+      return response!;
+    }
     return _petPost(body: body, cacheControl: cacheControl);
   }
 
   ///Add a new pet to the store
   ///@param body Pet object that needs to be added to the store
-  @Post(path: '/pet')
+  @chopper.Post(path: '/pet')
   Future<chopper.Response> _petPost({
     @Body() required Pet? body,
     @Header('Cache-Control') String? cacheControl,
@@ -65,15 +83,23 @@ abstract class PetServiceJson extends ChopperService {
   Future<chopper.Response> petPut({
     required Pet? body,
     String? cacheControl,
-  }) {
+  }) async {
     generatedMapping.putIfAbsent(Pet, () => Pet.fromJsonFactory);
 
+    if (_handleConnection != null) {
+      Future<chopper.Response>? response;
+      await _handleConnection!(() async {
+        response = _petPut(body: body, cacheControl: cacheControl);
+        await response;
+      });
+      return response!;
+    }
     return _petPut(body: body, cacheControl: cacheControl);
   }
 
   ///Update an existing pet
   ///@param body Pet object that needs to be added to the store
-  @Put(path: '/pet')
+  @chopper.Put(path: '/pet')
   Future<chopper.Response> _petPut({
     @Body() required Pet? body,
     @Header('Cache-Control') String? cacheControl,
@@ -84,16 +110,25 @@ abstract class PetServiceJson extends ChopperService {
   Future<chopper.Response<List<Pet>>> petFindByStatusGet({
     required enums.PetFindByStatusGetStatus? status,
     String? cacheControl,
-  }) {
+  }) async {
     generatedMapping.putIfAbsent(Pet, () => Pet.fromJsonFactory);
 
+    if (_handleConnection != null) {
+      Future<chopper.Response<List<Pet>>>? response;
+      await _handleConnection!(() async {
+        response = _petFindByStatusGet(
+            status: status?.value?.toString(), cacheControl: cacheControl);
+        await response;
+      });
+      return response!;
+    }
     return _petFindByStatusGet(
         status: status?.value?.toString(), cacheControl: cacheControl);
   }
 
   ///Finds Pets by status
   ///@param status Status values that need to be considered for filter
-  @Get(path: '/pet/findByStatus')
+  @chopper.Get(path: '/pet/findByStatus')
   Future<chopper.Response<List<Pet>>> _petFindByStatusGet({
     @Query('status') required String? status,
     @Header('Cache-Control') String? cacheControl,
@@ -104,15 +139,23 @@ abstract class PetServiceJson extends ChopperService {
   Future<chopper.Response<List<Pet>>> petFindByTagsGet({
     required List<String>? tags,
     String? cacheControl,
-  }) {
+  }) async {
     generatedMapping.putIfAbsent(Pet, () => Pet.fromJsonFactory);
 
+    if (_handleConnection != null) {
+      Future<chopper.Response<List<Pet>>>? response;
+      await _handleConnection!(() async {
+        response = _petFindByTagsGet(tags: tags, cacheControl: cacheControl);
+        await response;
+      });
+      return response!;
+    }
     return _petFindByTagsGet(tags: tags, cacheControl: cacheControl);
   }
 
   ///Finds Pets by tags
   ///@param tags Tags to filter by
-  @Get(path: '/pet/findByTags')
+  @chopper.Get(path: '/pet/findByTags')
   Future<chopper.Response<List<Pet>>> _petFindByTagsGet({
     @Query('tags') required List<String>? tags,
     @Header('Cache-Control') String? cacheControl,
@@ -122,21 +165,30 @@ abstract class PetServiceJson extends ChopperService {
   ///@param petId ID of pet to return
   Future<chopper.Response<Pet>> petPetIdGet({
     required int? petId,
-    String? apiKey,
+    ApiKey? apiKey,
     String? cacheControl,
-  }) {
+  }) async {
     generatedMapping.putIfAbsent(Pet, () => Pet.fromJsonFactory);
 
+    if (_handleConnection != null) {
+      Future<chopper.Response<Pet>>? response;
+      await _handleConnection!(() async {
+        response = _petPetIdGet(
+            petId: petId, apiKey: apiKey, cacheControl: cacheControl);
+        await response;
+      });
+      return response!;
+    }
     return _petPetIdGet(
         petId: petId, apiKey: apiKey, cacheControl: cacheControl);
   }
 
   ///Find pet by ID
   ///@param petId ID of pet to return
-  @Get(path: '/pet/{petId}')
+  @chopper.Get(path: '/pet/{petId}')
   Future<chopper.Response<Pet>> _petPetIdGet({
     @Path('petId') required int? petId,
-    @Header('api_key') String? apiKey,
+    @Header('api_key') ApiKey? apiKey,
     @Header('Cache-Control') String? cacheControl,
   });
 
@@ -149,7 +201,19 @@ abstract class PetServiceJson extends ChopperService {
     String? name,
     String? status,
     String? cacheControl,
-  }) {
+  }) async {
+    if (_handleConnection != null) {
+      Future<chopper.Response>? response;
+      await _handleConnection!(() async {
+        response = _petPetIdPost(
+            petId: petId,
+            name: name,
+            status: status,
+            cacheControl: cacheControl);
+        await response;
+      });
+      return response!;
+    }
     return _petPetIdPost(
         petId: petId, name: name, status: status, cacheControl: cacheControl);
   }
@@ -158,7 +222,7 @@ abstract class PetServiceJson extends ChopperService {
   ///@param petId ID of pet that needs to be updated
   ///@param name Updated name of the pet
   ///@param status Updated status of the pet
-  @Post(
+  @chopper.Post(
     path: '/pet/{petId}',
     optionalBody: true,
   )
@@ -176,7 +240,16 @@ abstract class PetServiceJson extends ChopperService {
     String? apiKey,
     required int? petId,
     String? cacheControl,
-  }) {
+  }) async {
+    if (_handleConnection != null) {
+      Future<chopper.Response>? response;
+      await _handleConnection!(() async {
+        response = _petPetIdDelete(
+            apiKey: apiKey, petId: petId, cacheControl: cacheControl);
+        await response;
+      });
+      return response!;
+    }
     return _petPetIdDelete(
         apiKey: apiKey, petId: petId, cacheControl: cacheControl);
   }
@@ -184,7 +257,7 @@ abstract class PetServiceJson extends ChopperService {
   ///Deletes a pet
   ///@param api_key
   ///@param petId Pet id to delete
-  @Delete(path: '/pet/{petId}')
+  @chopper.Delete(path: '/pet/{petId}')
   Future<chopper.Response> _petPetIdDelete({
     @Header('api_key') String? apiKey,
     @Path('petId') required int? petId,
@@ -200,10 +273,22 @@ abstract class PetServiceJson extends ChopperService {
     String? additionalMetadata,
     List<int>? file,
     String? cacheControl,
-  }) {
+  }) async {
     generatedMapping.putIfAbsent(
         ApiResponse, () => ApiResponse.fromJsonFactory);
 
+    if (_handleConnection != null) {
+      Future<chopper.Response<ApiResponse>>? response;
+      await _handleConnection!(() async {
+        response = _petPetIdUploadImagePost(
+            petId: petId,
+            additionalMetadata: additionalMetadata,
+            file: file,
+            cacheControl: cacheControl);
+        await response;
+      });
+      return response!;
+    }
     return _petPetIdUploadImagePost(
         petId: petId,
         additionalMetadata: additionalMetadata,
@@ -215,7 +300,7 @@ abstract class PetServiceJson extends ChopperService {
   ///@param petId ID of pet to update
   ///@param additionalMetadata Additional data to pass to server
   ///@param file file to upload
-  @Post(
+  @chopper.Post(
     path: '/pet/{petId}/uploadImage',
     optionalBody: true,
   )
@@ -228,16 +313,25 @@ abstract class PetServiceJson extends ChopperService {
 
   ///Returns pet inventories by status
   Future<chopper.Response<Object>> storeInventoryGet({
-    String? apiKey,
+    ApiKey? apiKey,
     String? cacheControl,
-  }) {
+  }) async {
+    if (_handleConnection != null) {
+      Future<chopper.Response<Object>>? response;
+      await _handleConnection!(() async {
+        response =
+            _storeInventoryGet(apiKey: apiKey, cacheControl: cacheControl);
+        await response;
+      });
+      return response!;
+    }
     return _storeInventoryGet(apiKey: apiKey, cacheControl: cacheControl);
   }
 
   ///Returns pet inventories by status
-  @Get(path: '/store/inventory')
+  @chopper.Get(path: '/store/inventory')
   Future<chopper.Response<Object>> _storeInventoryGet({
-    @Header('api_key') String? apiKey,
+    @Header('api_key') ApiKey? apiKey,
     @Header('Cache-Control') String? cacheControl,
   });
 
@@ -246,15 +340,23 @@ abstract class PetServiceJson extends ChopperService {
   Future<chopper.Response<Order>> storeOrderPost({
     required Order? body,
     String? cacheControl,
-  }) {
+  }) async {
     generatedMapping.putIfAbsent(Order, () => Order.fromJsonFactory);
 
+    if (_handleConnection != null) {
+      Future<chopper.Response<Order>>? response;
+      await _handleConnection!(() async {
+        response = _storeOrderPost(body: body, cacheControl: cacheControl);
+        await response;
+      });
+      return response!;
+    }
     return _storeOrderPost(body: body, cacheControl: cacheControl);
   }
 
   ///Place an order for a pet
   ///@param body order placed for purchasing the pet
-  @Post(path: '/store/order')
+  @chopper.Post(path: '/store/order')
   Future<chopper.Response<Order>> _storeOrderPost({
     @Body() required Order? body,
     @Header('Cache-Control') String? cacheControl,
@@ -265,15 +367,24 @@ abstract class PetServiceJson extends ChopperService {
   Future<chopper.Response<Order>> storeOrderOrderIdGet({
     required int? orderId,
     String? cacheControl,
-  }) {
+  }) async {
     generatedMapping.putIfAbsent(Order, () => Order.fromJsonFactory);
 
+    if (_handleConnection != null) {
+      Future<chopper.Response<Order>>? response;
+      await _handleConnection!(() async {
+        response =
+            _storeOrderOrderIdGet(orderId: orderId, cacheControl: cacheControl);
+        await response;
+      });
+      return response!;
+    }
     return _storeOrderOrderIdGet(orderId: orderId, cacheControl: cacheControl);
   }
 
   ///Find purchase order by ID
   ///@param orderId ID of pet that needs to be fetched
-  @Get(path: '/store/order/{orderId}')
+  @chopper.Get(path: '/store/order/{orderId}')
   Future<chopper.Response<Order>> _storeOrderOrderIdGet({
     @Path('orderId') required int? orderId,
     @Header('Cache-Control') String? cacheControl,
@@ -284,14 +395,23 @@ abstract class PetServiceJson extends ChopperService {
   Future<chopper.Response> storeOrderOrderIdDelete({
     required int? orderId,
     String? cacheControl,
-  }) {
+  }) async {
+    if (_handleConnection != null) {
+      Future<chopper.Response>? response;
+      await _handleConnection!(() async {
+        response = _storeOrderOrderIdDelete(
+            orderId: orderId, cacheControl: cacheControl);
+        await response;
+      });
+      return response!;
+    }
     return _storeOrderOrderIdDelete(
         orderId: orderId, cacheControl: cacheControl);
   }
 
   ///Delete purchase order by ID
   ///@param orderId ID of the order that needs to be deleted
-  @Delete(path: '/store/order/{orderId}')
+  @chopper.Delete(path: '/store/order/{orderId}')
   Future<chopper.Response> _storeOrderOrderIdDelete({
     @Path('orderId') required int? orderId,
     @Header('Cache-Control') String? cacheControl,
@@ -302,15 +422,23 @@ abstract class PetServiceJson extends ChopperService {
   Future<chopper.Response> userPost({
     required User? body,
     String? cacheControl,
-  }) {
+  }) async {
     generatedMapping.putIfAbsent(User, () => User.fromJsonFactory);
 
+    if (_handleConnection != null) {
+      Future<chopper.Response>? response;
+      await _handleConnection!(() async {
+        response = _userPost(body: body, cacheControl: cacheControl);
+        await response;
+      });
+      return response!;
+    }
     return _userPost(body: body, cacheControl: cacheControl);
   }
 
   ///Create user
   ///@param body Created user object
-  @Post(path: '/user')
+  @chopper.Post(path: '/user')
   Future<chopper.Response> _userPost({
     @Body() required User? body,
     @Header('Cache-Control') String? cacheControl,
@@ -321,13 +449,22 @@ abstract class PetServiceJson extends ChopperService {
   Future<chopper.Response> userCreateWithArrayPost({
     required List<User>? body,
     String? cacheControl,
-  }) {
+  }) async {
+    if (_handleConnection != null) {
+      Future<chopper.Response>? response;
+      await _handleConnection!(() async {
+        response =
+            _userCreateWithArrayPost(body: body, cacheControl: cacheControl);
+        await response;
+      });
+      return response!;
+    }
     return _userCreateWithArrayPost(body: body, cacheControl: cacheControl);
   }
 
   ///Creates list of users with given input array
   ///@param body List of user object
-  @Post(path: '/user/createWithArray')
+  @chopper.Post(path: '/user/createWithArray')
   Future<chopper.Response> _userCreateWithArrayPost({
     @Body() required List<User>? body,
     @Header('Cache-Control') String? cacheControl,
@@ -338,13 +475,22 @@ abstract class PetServiceJson extends ChopperService {
   Future<chopper.Response> userCreateWithListPost({
     required List<User>? body,
     String? cacheControl,
-  }) {
+  }) async {
+    if (_handleConnection != null) {
+      Future<chopper.Response>? response;
+      await _handleConnection!(() async {
+        response =
+            _userCreateWithListPost(body: body, cacheControl: cacheControl);
+        await response;
+      });
+      return response!;
+    }
     return _userCreateWithListPost(body: body, cacheControl: cacheControl);
   }
 
   ///Creates list of users with given input array
   ///@param body List of user object
-  @Post(path: '/user/createWithList')
+  @chopper.Post(path: '/user/createWithList')
   Future<chopper.Response> _userCreateWithListPost({
     @Body() required List<User>? body,
     @Header('Cache-Control') String? cacheControl,
@@ -357,7 +503,16 @@ abstract class PetServiceJson extends ChopperService {
     required String? username,
     required String? password,
     String? cacheControl,
-  }) {
+  }) async {
+    if (_handleConnection != null) {
+      Future<chopper.Response<String>>? response;
+      await _handleConnection!(() async {
+        response = _userLoginGet(
+            username: username, password: password, cacheControl: cacheControl);
+        await response;
+      });
+      return response!;
+    }
     return _userLoginGet(
         username: username, password: password, cacheControl: cacheControl);
   }
@@ -365,7 +520,7 @@ abstract class PetServiceJson extends ChopperService {
   ///Logs user into the system
   ///@param username The user name for login
   ///@param password The password for login in clear text
-  @Get(path: '/user/login')
+  @chopper.Get(path: '/user/login')
   Future<chopper.Response<String>> _userLoginGet({
     @Query('username') required String? username,
     @Query('password') required String? password,
@@ -373,12 +528,20 @@ abstract class PetServiceJson extends ChopperService {
   });
 
   ///Logs out current logged in user session
-  Future<chopper.Response> userLogoutGet({String? cacheControl}) {
+  Future<chopper.Response> userLogoutGet({String? cacheControl}) async {
+    if (_handleConnection != null) {
+      Future<chopper.Response>? response;
+      await _handleConnection!(() async {
+        response = _userLogoutGet(cacheControl: cacheControl);
+        await response;
+      });
+      return response!;
+    }
     return _userLogoutGet(cacheControl: cacheControl);
   }
 
   ///Logs out current logged in user session
-  @Get(path: '/user/logout')
+  @chopper.Get(path: '/user/logout')
   Future<chopper.Response> _userLogoutGet(
       {@Header('Cache-Control') String? cacheControl});
 
@@ -387,15 +550,24 @@ abstract class PetServiceJson extends ChopperService {
   Future<chopper.Response<User>> userUsernameGet({
     required String? username,
     String? cacheControl,
-  }) {
+  }) async {
     generatedMapping.putIfAbsent(User, () => User.fromJsonFactory);
 
+    if (_handleConnection != null) {
+      Future<chopper.Response<User>>? response;
+      await _handleConnection!(() async {
+        response =
+            _userUsernameGet(username: username, cacheControl: cacheControl);
+        await response;
+      });
+      return response!;
+    }
     return _userUsernameGet(username: username, cacheControl: cacheControl);
   }
 
   ///Get user by user name
   ///@param username The name that needs to be fetched. Use user1 for testing.
-  @Get(path: '/user/{username}')
+  @chopper.Get(path: '/user/{username}')
   Future<chopper.Response<User>> _userUsernameGet({
     @Path('username') required String? username,
     @Header('Cache-Control') String? cacheControl,
@@ -408,9 +580,18 @@ abstract class PetServiceJson extends ChopperService {
     required String? username,
     required User? body,
     String? cacheControl,
-  }) {
+  }) async {
     generatedMapping.putIfAbsent(User, () => User.fromJsonFactory);
 
+    if (_handleConnection != null) {
+      Future<chopper.Response>? response;
+      await _handleConnection!(() async {
+        response = _userUsernamePut(
+            username: username, body: body, cacheControl: cacheControl);
+        await response;
+      });
+      return response!;
+    }
     return _userUsernamePut(
         username: username, body: body, cacheControl: cacheControl);
   }
@@ -418,7 +599,7 @@ abstract class PetServiceJson extends ChopperService {
   ///Updated user
   ///@param username name that need to be updated
   ///@param body Updated user object
-  @Put(path: '/user/{username}')
+  @chopper.Put(path: '/user/{username}')
   Future<chopper.Response> _userUsernamePut({
     @Path('username') required String? username,
     @Body() required User? body,
@@ -430,20 +611,154 @@ abstract class PetServiceJson extends ChopperService {
   Future<chopper.Response> userUsernameDelete({
     required String? username,
     String? cacheControl,
-  }) {
+  }) async {
+    if (_handleConnection != null) {
+      Future<chopper.Response>? response;
+      await _handleConnection!(() async {
+        response =
+            _userUsernameDelete(username: username, cacheControl: cacheControl);
+        await response;
+      });
+      return response!;
+    }
     return _userUsernameDelete(username: username, cacheControl: cacheControl);
   }
 
   ///Delete user
   ///@param username The name that needs to be deleted
-  @Delete(path: '/user/{username}')
+  @chopper.Delete(path: '/user/{username}')
   Future<chopper.Response> _userUsernameDelete({
     @Path('username') required String? username,
     @Header('Cache-Control') String? cacheControl,
   });
 }
 
+class LocalDateTimeConverter implements json.JsonConverter<DateTime, String> {
+  const LocalDateTimeConverter();
+
+  @override
+  DateTime fromJson(String json) => DateTime.parse(json).toLocal();
+
+  @override
+  String toJson(DateTime object) => object.toUtc().toString();
+}
+
+class OpenApiException implements Exception {
+  const OpenApiException(this.error);
+
+  final Object error;
+
+  @override
+  String toString() => 'OpenApiException: ${error.toString()}';
+}
+
 @JsonSerializable(explicitToJson: true)
+@LocalDateTimeConverter()
+class Order {
+  Order({
+    this.id,
+    this.petId,
+    this.quantity,
+    this.shipDate,
+    this.status,
+    this.complete,
+  });
+
+  factory Order.fromJson(Map<String, dynamic> json) => _$OrderFromJson(json);
+
+  static const toJsonFactory = _$OrderToJson;
+  Map<String, dynamic> toJson() => _$OrderToJson(this);
+
+  @JsonKey(name: 'id', includeIfNull: false)
+  int? id;
+  @JsonKey(name: 'petId', includeIfNull: false)
+  int? petId;
+  @JsonKey(name: 'quantity', includeIfNull: false)
+  int? quantity;
+  @JsonKey(name: 'shipDate', includeIfNull: false)
+  DateTime? shipDate;
+  @JsonKey(
+    name: 'status',
+    includeIfNull: false,
+    toJson: orderStatusToJson,
+    fromJson: orderStatusFromJson,
+  )
+  enums.OrderStatus? status;
+  @JsonKey(name: 'complete', includeIfNull: false, defaultValue: false)
+  bool? complete;
+  static const fromJsonFactory = _$OrderFromJson;
+
+  @override
+  bool operator ==(dynamic other) {
+    return identical(this, other) ||
+        (other is Order &&
+            (identical(other.id, id) ||
+                const DeepCollectionEquality().equals(other.id, id)) &&
+            (identical(other.petId, petId) ||
+                const DeepCollectionEquality().equals(other.petId, petId)) &&
+            (identical(other.quantity, quantity) ||
+                const DeepCollectionEquality()
+                    .equals(other.quantity, quantity)) &&
+            (identical(other.shipDate, shipDate) ||
+                const DeepCollectionEquality()
+                    .equals(other.shipDate, shipDate)) &&
+            (identical(other.status, status) ||
+                const DeepCollectionEquality().equals(other.status, status)) &&
+            (identical(other.complete, complete) ||
+                const DeepCollectionEquality()
+                    .equals(other.complete, complete)));
+  }
+
+  @override
+  String toString() => jsonEncode(this);
+
+  @override
+  int get hashCode =>
+      const DeepCollectionEquality().hash(id) ^
+      const DeepCollectionEquality().hash(petId) ^
+      const DeepCollectionEquality().hash(quantity) ^
+      const DeepCollectionEquality().hash(shipDate) ^
+      const DeepCollectionEquality().hash(status) ^
+      const DeepCollectionEquality().hash(complete) ^
+      runtimeType.hashCode;
+}
+
+extension $OrderExtension on Order {
+  Order copyWith(
+      {int? id,
+      int? petId,
+      int? quantity,
+      DateTime? shipDate,
+      enums.OrderStatus? status,
+      bool? complete}) {
+    return Order(
+        id: id ?? this.id,
+        petId: petId ?? this.petId,
+        quantity: quantity ?? this.quantity,
+        shipDate: shipDate ?? this.shipDate,
+        status: status ?? this.status,
+        complete: complete ?? this.complete);
+  }
+
+  Order copyWithWrapped(
+      {Wrapped<int?>? id,
+      Wrapped<int?>? petId,
+      Wrapped<int?>? quantity,
+      Wrapped<DateTime?>? shipDate,
+      Wrapped<enums.OrderStatus?>? status,
+      Wrapped<bool?>? complete}) {
+    return Order(
+        id: (id != null ? id.value : this.id),
+        petId: (petId != null ? petId.value : this.petId),
+        quantity: (quantity != null ? quantity.value : this.quantity),
+        shipDate: (shipDate != null ? shipDate.value : this.shipDate),
+        status: (status != null ? status.value : this.status),
+        complete: (complete != null ? complete.value : this.complete));
+  }
+}
+
+@JsonSerializable(explicitToJson: true)
+@LocalDateTimeConverter()
 class Category {
   Category({
     this.id,
@@ -453,13 +768,14 @@ class Category {
   factory Category.fromJson(Map<String, dynamic> json) =>
       _$CategoryFromJson(json);
 
-  @JsonKey(name: 'id', includeIfNull: false)
-  final num? id;
-  @JsonKey(name: 'name', includeIfNull: false, defaultValue: '')
-  final String? name;
-  static const fromJsonFactory = _$CategoryFromJson;
   static const toJsonFactory = _$CategoryToJson;
   Map<String, dynamic> toJson() => _$CategoryToJson(this);
+
+  @JsonKey(name: 'id', includeIfNull: false)
+  int? id;
+  @JsonKey(name: 'name', includeIfNull: false, defaultValue: '')
+  String? name;
+  static const fromJsonFactory = _$CategoryFromJson;
 
   @override
   bool operator ==(dynamic other) {
@@ -482,11 +798,11 @@ class Category {
 }
 
 extension $CategoryExtension on Category {
-  Category copyWith({num? id, String? name}) {
+  Category copyWith({int? id, String? name}) {
     return Category(id: id ?? this.id, name: name ?? this.name);
   }
 
-  Category copyWithWrapped({Wrapped<num?>? id, Wrapped<String?>? name}) {
+  Category copyWithWrapped({Wrapped<int?>? id, Wrapped<String?>? name}) {
     return Category(
         id: (id != null ? id.value : this.id),
         name: (name != null ? name.value : this.name));
@@ -494,6 +810,7 @@ extension $CategoryExtension on Category {
 }
 
 @JsonSerializable(explicitToJson: true)
+@LocalDateTimeConverter()
 class User {
   User({
     this.id,
@@ -508,25 +825,26 @@ class User {
 
   factory User.fromJson(Map<String, dynamic> json) => _$UserFromJson(json);
 
-  @JsonKey(name: 'id', includeIfNull: false)
-  final num? id;
-  @JsonKey(name: 'username', includeIfNull: false, defaultValue: '')
-  final String? username;
-  @JsonKey(name: 'firstName', includeIfNull: false, defaultValue: '')
-  final String? firstName;
-  @JsonKey(name: 'lastName', includeIfNull: false, defaultValue: '')
-  final String? lastName;
-  @JsonKey(name: 'email', includeIfNull: false, defaultValue: '')
-  final String? email;
-  @JsonKey(name: 'password', includeIfNull: false, defaultValue: '')
-  final String? password;
-  @JsonKey(name: 'phone', includeIfNull: false, defaultValue: '')
-  final String? phone;
-  @JsonKey(name: 'userStatus', includeIfNull: false)
-  final int? userStatus;
-  static const fromJsonFactory = _$UserFromJson;
   static const toJsonFactory = _$UserToJson;
   Map<String, dynamic> toJson() => _$UserToJson(this);
+
+  @JsonKey(name: 'id', includeIfNull: false)
+  int? id;
+  @JsonKey(name: 'username', includeIfNull: false, defaultValue: '')
+  String? username;
+  @JsonKey(name: 'firstName', includeIfNull: false, defaultValue: '')
+  String? firstName;
+  @JsonKey(name: 'lastName', includeIfNull: false, defaultValue: '')
+  String? lastName;
+  @JsonKey(name: 'email', includeIfNull: false, defaultValue: '')
+  String? email;
+  @JsonKey(name: 'password', includeIfNull: false, defaultValue: '')
+  String? password;
+  @JsonKey(name: 'phone', includeIfNull: false, defaultValue: '')
+  String? phone;
+  @JsonKey(name: 'userStatus', includeIfNull: false)
+  int? userStatus;
+  static const fromJsonFactory = _$UserFromJson;
 
   @override
   bool operator ==(dynamic other) {
@@ -573,7 +891,7 @@ class User {
 
 extension $UserExtension on User {
   User copyWith(
-      {num? id,
+      {int? id,
       String? username,
       String? firstName,
       String? lastName,
@@ -593,7 +911,7 @@ extension $UserExtension on User {
   }
 
   User copyWithWrapped(
-      {Wrapped<num?>? id,
+      {Wrapped<int?>? id,
       Wrapped<String?>? username,
       Wrapped<String?>? firstName,
       Wrapped<String?>? lastName,
@@ -614,6 +932,7 @@ extension $UserExtension on User {
 }
 
 @JsonSerializable(explicitToJson: true)
+@LocalDateTimeConverter()
 class Tag {
   Tag({
     this.id,
@@ -622,13 +941,14 @@ class Tag {
 
   factory Tag.fromJson(Map<String, dynamic> json) => _$TagFromJson(json);
 
-  @JsonKey(name: 'id', includeIfNull: false)
-  final num? id;
-  @JsonKey(name: 'name', includeIfNull: false, defaultValue: '')
-  final String? name;
-  static const fromJsonFactory = _$TagFromJson;
   static const toJsonFactory = _$TagToJson;
   Map<String, dynamic> toJson() => _$TagToJson(this);
+
+  @JsonKey(name: 'id', includeIfNull: false)
+  int? id;
+  @JsonKey(name: 'name', includeIfNull: false, defaultValue: '')
+  String? name;
+  static const fromJsonFactory = _$TagFromJson;
 
   @override
   bool operator ==(dynamic other) {
@@ -651,11 +971,11 @@ class Tag {
 }
 
 extension $TagExtension on Tag {
-  Tag copyWith({num? id, String? name}) {
+  Tag copyWith({int? id, String? name}) {
     return Tag(id: id ?? this.id, name: name ?? this.name);
   }
 
-  Tag copyWithWrapped({Wrapped<num?>? id, Wrapped<String?>? name}) {
+  Tag copyWithWrapped({Wrapped<int?>? id, Wrapped<String?>? name}) {
     return Tag(
         id: (id != null ? id.value : this.id),
         name: (name != null ? name.value : this.name));
@@ -663,6 +983,7 @@ extension $TagExtension on Tag {
 }
 
 @JsonSerializable(explicitToJson: true)
+@LocalDateTimeConverter()
 class Pet {
   Pet({
     this.id,
@@ -675,26 +996,27 @@ class Pet {
 
   factory Pet.fromJson(Map<String, dynamic> json) => _$PetFromJson(json);
 
+  static const toJsonFactory = _$PetToJson;
+  Map<String, dynamic> toJson() => _$PetToJson(this);
+
   @JsonKey(name: 'id', includeIfNull: false)
-  final num? id;
+  int? id;
   @JsonKey(name: 'category', includeIfNull: false)
-  final Category? category;
+  Category? category;
   @JsonKey(name: 'name', includeIfNull: false, defaultValue: '')
-  final String? name;
+  String? name;
   @JsonKey(name: 'photoUrls', includeIfNull: false)
-  final List<String>? photoUrls;
+  List<String>? photoUrls;
   @JsonKey(name: 'tags', includeIfNull: false)
-  final List<Tag>? tags;
+  List<Tag>? tags;
   @JsonKey(
     name: 'status',
     includeIfNull: false,
     toJson: petStatusToJson,
     fromJson: petStatusFromJson,
   )
-  final enums.PetStatus? status;
+  enums.PetStatus? status;
   static const fromJsonFactory = _$PetFromJson;
-  static const toJsonFactory = _$PetToJson;
-  Map<String, dynamic> toJson() => _$PetToJson(this);
 
   @override
   bool operator ==(dynamic other) {
@@ -732,7 +1054,7 @@ class Pet {
 
 extension $PetExtension on Pet {
   Pet copyWith(
-      {num? id,
+      {int? id,
       Category? category,
       String? name,
       List<String>? photoUrls,
@@ -748,7 +1070,7 @@ extension $PetExtension on Pet {
   }
 
   Pet copyWithWrapped(
-      {Wrapped<num?>? id,
+      {Wrapped<int?>? id,
       Wrapped<Category?>? category,
       Wrapped<String?>? name,
       Wrapped<List<String>?>? photoUrls,
@@ -765,6 +1087,7 @@ extension $PetExtension on Pet {
 }
 
 @JsonSerializable(explicitToJson: true)
+@LocalDateTimeConverter()
 class ApiResponse {
   ApiResponse({
     this.code,
@@ -775,15 +1098,16 @@ class ApiResponse {
   factory ApiResponse.fromJson(Map<String, dynamic> json) =>
       _$ApiResponseFromJson(json);
 
-  @JsonKey(name: 'code', includeIfNull: false)
-  final int? code;
-  @JsonKey(name: 'type', includeIfNull: false, defaultValue: '')
-  final String? type;
-  @JsonKey(name: 'message', includeIfNull: false, defaultValue: '')
-  final String? message;
-  static const fromJsonFactory = _$ApiResponseFromJson;
   static const toJsonFactory = _$ApiResponseToJson;
   Map<String, dynamic> toJson() => _$ApiResponseToJson(this);
+
+  @JsonKey(name: 'code', includeIfNull: false)
+  int? code;
+  @JsonKey(name: 'type', includeIfNull: false, defaultValue: '')
+  String? type;
+  @JsonKey(name: 'message', includeIfNull: false, defaultValue: '')
+  String? message;
+  static const fromJsonFactory = _$ApiResponseFromJson;
 
   @override
   bool operator ==(dynamic other) {
@@ -835,8 +1159,9 @@ enums.OrderStatus orderStatusFromJson(
   Object? orderStatus, [
   enums.OrderStatus? defaultValue,
 ]) {
-  return enums.OrderStatus.values
-          .firstWhereOrNull((e) => e.value == orderStatus) ??
+  return enums.OrderStatus.values.firstWhereOrNull((e) =>
+          e.value.toString().toLowerCase() ==
+          orderStatus?.toString()?.toLowerCase()) ??
       defaultValue ??
       enums.OrderStatus.swaggerGeneratedUnknown;
 }
@@ -879,7 +1204,9 @@ enums.PetStatus petStatusFromJson(
   Object? petStatus, [
   enums.PetStatus? defaultValue,
 ]) {
-  return enums.PetStatus.values.firstWhereOrNull((e) => e.value == petStatus) ??
+  return enums.PetStatus.values.firstWhereOrNull((e) =>
+          e.value.toString().toLowerCase() ==
+          petStatus?.toString()?.toLowerCase()) ??
       defaultValue ??
       enums.PetStatus.swaggerGeneratedUnknown;
 }
@@ -923,8 +1250,9 @@ enums.PetFindByStatusGetStatus petFindByStatusGetStatusFromJson(
   Object? petFindByStatusGetStatus, [
   enums.PetFindByStatusGetStatus? defaultValue,
 ]) {
-  return enums.PetFindByStatusGetStatus.values
-          .firstWhereOrNull((e) => e.value == petFindByStatusGetStatus) ??
+  return enums.PetFindByStatusGetStatus.values.firstWhereOrNull((e) =>
+          e.value.toString().toLowerCase() ==
+          petFindByStatusGetStatus?.toString()?.toLowerCase()) ??
       defaultValue ??
       enums.PetFindByStatusGetStatus.swaggerGeneratedUnknown;
 }