+++ lib/src/code_generators/swagger_models_generator.dart
@@ -62,8 +62,12 @@ abstract class SwaggerModelsGenerator extends SwaggerGeneratorBase {
       return 'class $className {}';
     }
 
-    if (schema.type == kObject && schema.anyOf.isNotEmpty) {
-      return 'typedef $className = Map<String, dynamic>;';
+    if (schema.anyOf.isNotEmpty) {
+      if (schema.type == kObject) {
+        return 'typedef $className = Map<String, dynamic>;';
+      } else {
+        return 'typedef $className = Object;';
+      }
     }
 
     if (schema.type == 'array') {
@@ -138,11 +142,28 @@ abstract class SwaggerModelsGenerator extends SwaggerGeneratorBase {
     final result = <String, SwaggerSchema>{};
 
     classes.forEach((classKey, schema) {
-      final properties = schema.properties;
+      final properties = {
+        ...schema.properties,
+        ...schema.items?.properties ?? {},
+      };
+
+      for (var element in schema.allOf) {
+        properties.addAll(element.properties);
+
+        if (element.ref.isNotEmpty) {
+          final neededClass = classes[element.ref.getUnformattedRef()];
+          properties.addAll(neededClass?.properties ?? {});
+        }
+      }
+
+      final shouldUseItemsProperties =
+          schema.items?.properties.isNotEmpty == true;
 
       properties.forEach((propertyKey, propSchema) {
+        final itemPart = shouldUseItemsProperties ? '\$Item' : '';
+
         final innerClassName =
-            '${getValidatedClassName(classKey)}\$${getValidatedClassName(propertyKey)}';
+            '${getValidatedClassName(classKey)}$itemPart\$${getValidatedClassName(propertyKey)}';
 
         if (propSchema.properties.isNotEmpty) {
           result[innerClassName] = propSchema;
@@ -151,9 +172,15 @@ abstract class SwaggerModelsGenerator extends SwaggerGeneratorBase {
         final items = propSchema.items;
 
         if (items != null && items.properties.isNotEmpty) {
-          result[innerClassName] = propSchema;
+          propSchema.type = 'object';
+          result['$innerClassName\$Item'] = items;
         }
       });
+
+      if (schema.items != null) {
+        result.addAll(getClassesFromInnerClasses(
+            {'${getValidatedClassName(classKey)}\$item': schema.items!}));
+      }
     });
 
     if (result.isNotEmpty) {
@@ -224,7 +251,9 @@ abstract class SwaggerModelsGenerator extends SwaggerGeneratorBase {
     required bool generateEnumsMethods,
   }) {
     final allEnumsString = generateEnumsMethods
-        ? allEnums.map((e) => e.generateFromJsonToJson()).join()
+        ? allEnums
+            .map((e) => e.generateFromJsonToJson(options.enumsCaseSensitive))
+            .join()
         : '';
 
     final allEnumListNames = getAllListEnumNames(root);
@@ -321,6 +350,10 @@ abstract class SwaggerModelsGenerator extends SwaggerGeneratorBase {
 
     if (parameter.properties.isNotEmpty) {
       return '${getValidatedClassName(className)}\$${getValidatedClassName(parameterName)}$modelPostfix';
+    }
+
+    if (parameter.items?.properties.isNotEmpty == true) {
+      return 'List<${getValidatedClassName(className)}\$${getValidatedClassName(parameterName)}\$Item$modelPostfix>';
     }
 
     if (parameter.hasRef) {
@@ -600,7 +633,11 @@ static $returnType $fromJsonFunction($valueType? value) => $enumNameCamelCase$fr
     final allOf = prop.allOf;
     String typeName;
 
-    if (allOf.length != 1) {
+    if (allOf
+            .where((element) =>
+                element.ref.isNotEmpty || element.properties.isNotEmpty)
+            .length >
+        1) {
       typeName = kDynamic;
     } else {
       var className = allOf.first.ref.getRef();
@@ -1045,7 +1082,11 @@ static $returnType $fromJsonFunction($valueType? value) => $enumNameCamelCase$fr
 
       final basicTypesMap = generateBasicTypesMapFromSchemas(root);
 
-      propertyName = propertyName.asParameterName();
+      propertyName = getValidatedParameterName(propertyName).asParameterName();
+
+      if (propertyName.isEmpty) {
+        propertyName = '\$';
+      }
 
       propertyName = getParameterName(propertyName, propertyNames);
 
@@ -1199,7 +1240,8 @@ static $returnType $fromJsonFunction($valueType? value) => $enumNameCamelCase$fr
 
     entityMap.forEach((key, value) {
       var fieldName = generateFieldName(
-        getParameterName(key.asParameterName(), propertyNames),
+        getParameterName(
+            getValidatedParameterName(key).asParameterName(), propertyNames),
       );
 
       propertyNames.add(fieldName);