+++ lib/src/code_generators/swagger_requests_generator.dart
@@ -92,8 +92,20 @@ class SwaggerRequestsGenerator extends SwaggerGeneratorBase {
         ..optionalParameters.add(Parameter(
           (p) => p
             ..named = true
+            ..type = Reference('http.Client?')
+            ..name = 'httpClient',
+        ))
+        ..optionalParameters.add(Parameter(
+          (p) => p
+            ..named = true
             ..type = Reference('Authenticator?')
             ..name = 'authenticator',
+        ))
+        ..optionalParameters.add(Parameter(
+          (p) => p
+            ..named = true
+            ..type = Reference('Converter?')
+            ..name = 'converter',
         ))
         ..optionalParameters.add(Parameter(
           (p) => p
@@ -140,6 +152,10 @@ class SwaggerRequestsGenerator extends SwaggerGeneratorBase {
             !options.includePaths
                 .any((includePath) => RegExp(includePath).hasMatch(path))) {
           return;
+        }
+
+        if (options.addBasePathToRequests) {
+          path = '${swaggerRoot.basePath}$path';
         }
 
         final methodName = _getRequestMethodName(
@@ -570,10 +586,9 @@ return _$publicMethodName($parametersListString);
   }) {
     final format = parameter.schema?.format ?? '';
 
-    if (parameter.inParameter == kHeader) {
-      return _mapParameterName(kString, format, '');
-    } else if (parameter.items?.enumValues.isNotEmpty == true ||
-        parameter.schema?.enumValues.isNotEmpty == true) {
+    if (parameter.items?.enumValues.isNotEmpty == true ||
+        parameter.schema?.enumValues.isNotEmpty == true ||
+        parameter.enumValues.isNotEmpty) {
       if (definedParameters.containsValue(parameter)) {
         return getValidatedClassName(parameter.name).asEnum();
       }
@@ -769,20 +784,28 @@ return _$publicMethodName($parametersListString);
 
         // otherwise no request scheme is defined, we provide every param as a separate param.
         schema?.properties.forEach((key, value) {
-          if (value.type == 'string' && value.format == 'binary') {
-            final isRequired = schema!.required.contains(key);
+          isBinary(SwaggerSchema? value) =>
+              (value?.type == 'string' && value?.format == 'binary') ||
+              value?.type == 'file';
+          if ((isBinary(value) ||
+              value.type == 'array' && isBinary(value.items))) {
+            final isRequired =
+                value.type == 'array' || schema!.required.contains(key);
+            String typeRef = isRequired
+                ? options.multipartFileType
+                : options.multipartFileType.makeNullable();
+
+            if (value.type == 'array') {
+              typeRef = 'List<$typeRef>';
+            }
+
             result.add(
               Parameter(
                 (p) => p
                   ..name = key
                   ..named = true
                   ..required = isRequired
-                  ..type = Reference(
-                    isRequired
-                        ? options.multipartFileType
-                        : options.multipartFileType.makeNullable(),
-                  )
-                  ..named = true
+                  ..type = Reference(typeRef)
                   ..annotations.add(
                     refer(kPartFile.pascalCase).call([]),
                   ),
@@ -1249,8 +1272,8 @@ return _$publicMethodName($parametersListString);
         : 'baseUrl: baseUrl';
 
     final converterString = options.withConverter
-        ? 'converter: \$JsonSerializableConverter(),'
-        : 'converter: chopper.JsonConverter(),';
+        ? 'converter: converter ?? \$JsonSerializableConverter(),'
+        : 'converter: converter ?? chopper.JsonConverter(),';
 
     final chopperClientBody = '''
     _handleConnection = handleConnection;
@@ -1263,6 +1286,7 @@ return _$publicMethodName($parametersListString);
       services: [_\$$className()],
       $converterString
       interceptors: interceptors ?? [],
+      client: httpClient,
       authenticator: authenticator,
       $baseUrlString);
     return _\$$className(newClient);