+++ test/generator_tests/enums_generator_test.dart
@@ -4,6 +4,7 @@ import 'package:swagger_dart_code_generator/src/models/generator_options.dart';
 import 'package:swagger_dart_code_generator/src/swagger_models/requests/swagger_request_parameter.dart';
 import 'package:test/test.dart';
 
+
 void main() {
   final generator = SwaggerEnumsGeneratorV3(
     GeneratorOptions(
@@ -46,7 +47,7 @@ void main() {
           SwaggerRequestParameter(ref: '#definitions/TestParameter');
       final overriddenParameters = [
         SwaggerRequestParameter(
-            name: 'TestParameter', type: 'Overridden parameter')
+            key: 'TestParameter', type: 'Overridden parameter')
       ];
 
       final result = generator.getOriginalOrOverriddenRequestParameter(