+++ test/generator_tests/models_generator_test.dart
@@ -532,6 +532,19 @@ void main() {
     });
   });
 
+  group('Tests for generateResponses', () {
+    test('Should generate empty string for V2', () {
+      final map = SwaggerRoot.parse(schemasResponsesWithResponse);
+      final result = generator2.generate(
+        root: map,
+        fileName: 'fileName',
+        allEnums: [],
+      );
+
+      expect(result, equals(''));
+    });
+  });
+
   group('Tests for models from responses', () {
     test('Should generate correct model from response', () {
       final map = SwaggerRoot.parse(requestWithReturnTypeInjected);